IEnumerable: Use yield for large size results
Teach Permutation(), Combination(), and Scrabble() methods to work
for results with large size by using yield.

With items cached,

    var atoz = "abcdefghijklmnopqrstuvwxyz"
    atoz.Combine(25)    // returns 26 items

would take ~2s, but takes 15s-40s with iterator.

But with iterator,

    atoz.Permute(8)     // returns 62,990,928,000 items

is possible and not even possible with items cached.

For Permututation, iterator requires O(r) space, and items cached
nPr space.

So err on the side of iterator.