Deque: Cover Delete(node) corner cases
Null out the node's owner in Delete(node) to cover corner case of
double-deleting a node and after dequeuing a node.

This does not cover Clear(). A guid will cover all though by
generating a new guid on Clear(), and nulling out owner on Delete(node)
and Dequeue(). But 16B every node is a lot of overhead so just let
Clear() be.

Misc:
 - Throw EmptyException if empty instead of InvalidOperation.